+++ brian2/monitors/spikemonitor.py
@@ -3,6 +3,7 @@ import numbers
 import numpy as np
 
 from brian2.core.variables import Variables
+from brian2.core.names import Nameable
 from brian2.units.allunits import second
 from brian2.units.fundamentalunits import Unit, Quantity
 from brian2.groups.group import CodeRunner, Group
@@ -27,12 +28,18 @@ class EventMonitor(Group, CodeRunner):
     record : bool
         Whether or not to record each event in `i` and `t` (the `count` will
         always be recorded).
+    event : str
+        The name of the event to record
+    variables : str or sequence of str, optional
+        Which variables to record at the time of the event (in addition to the
+        index of the neuron). Can be the name of a variable or a list of names.
     when : str, optional
-        When to record the events, by default records events in the slot
-        ``'end'``.
+        When to record the events, by default records events in the same slot
+        where the event is emitted.
     order : int, optional
         The priority of of this group for operations occurring at the same time
-        step and in the same scheduling slot. Defaults to 0.
+        step and in the same scheduling slot. Defaults to the order where the
+        event is emitted + 1, i.e. it will be recorded directly afterwards.
     name : str, optional
         A unique name for the object, otherwise will use
         ``source.name+'_eventmonitor_0'``, etc.
@@ -46,31 +53,57 @@ class EventMonitor(Group, CodeRunner):
     invalidates_magic_network = False
     add_to_magic_network = True
 
-    def __init__(self, source, event, when='end', order=0,
+    def __init__(self, source, event, variables=None, when=None, order=None,
                  name='eventmonitor*', codeobj_class=None):
         #: The source we are recording from
         self.source = source
 
+        if when is None:
+            if order is not None:
+                raise ValueError('Cannot specify order if when is not specified.')
+            if hasattr(source, 'thresholder'):
+                parent_obj = source.thresholder[event]
+            else:
+                parent_obj = source
+            when = parent_obj.when
+            order = parent_obj.order + 1
+        elif order is None:
+            order = 0
+
         #: The event that we are listening to
         self.event = event
 
+        if variables is None:
+            variables = []
+        elif isinstance(variables, basestring):
+            variables = [variables]
+
+        #: The additional variables that will be recorded
+        self.record_variables = variables
+
+        for variable in variables:
+            if variable not in source.variables:
+                raise ValueError(("'%s' is not a variable of the recorded "
+                                  "group" % variable))
+
+        # Some dummy code so that code generation takes care of the indexing
+        # and subexpressions
+        code = ['_to_record_%s = _source_%s' % (v, v)
+                for v in self.record_variables]
+        code = '\n'.join(code)
+
         self.codeobj_class = codeobj_class
+
         # Since this now works for general events not only spikes, we have to
         # pass the information about which variable to use to the template,
         # it can not longer simply refer to "_spikespace"
         eventspace_name = '_{}space'.format(event)
-        template_kwds = {'eventspace_variable': source.variables[eventspace_name]}
-        needed_variables= [eventspace_name]
-        CodeRunner.__init__(self, group=self, code='', template='spikemonitor',
-                            name=name, clock=source.clock, when=when,
-                            order=order, needed_variables=needed_variables,
-                            template_kwds=template_kwds)
-
-        self.add_dependency(source)
 
         # Handle subgroups correctly
         start = getattr(source, 'start', 0)
         stop = getattr(source, 'stop', len(source))
+
+        Nameable.__init__(self, name=name)
 
         self.variables = Variables(self)
         self.variables.add_reference(eventspace_name, source)
@@ -78,6 +111,17 @@ class EventMonitor(Group, CodeRunner):
                                          dtype=np.int32, constant_size=False)
         self.variables.add_dynamic_array('t', size=0, unit=second,
                                          constant_size=False)
+        for variable in self.record_variables:
+            source_var = source.variables[variable]
+            self.variables.add_reference('_source_%s' % variable,
+                                         source, variable)
+            self.variables.add_auxiliary_variable('_to_record_%s' % variable,
+                                                   unit=source_var.unit,
+                                                   dtype=source_var.dtype)
+            self.variables.add_dynamic_array(variable, size=0,
+                                             unit=source_var.unit,
+                                             dtype=source_var.dtype,
+                                             constant_size=False)
         self.variables.add_arange('_source_i', size=len(source))
         self.variables.add_array('count', size=len(source), unit=Unit(1),
                                  dtype=np.int32, read_only=True,
@@ -86,8 +130,22 @@ class EventMonitor(Group, CodeRunner):
         self.variables.add_constant('_source_stop', Unit(1), stop)
         self.variables.add_attribute_variable('N', unit=Unit(1), obj=self,
                                               attribute='_N', dtype=np.int32)
+
+        record_variables = {varname: self.variables[varname]
+                            for varname in self.record_variables}
+        template_kwds = {'eventspace_variable': source.variables[eventspace_name],
+                         'record_variables': record_variables}
+        needed_variables = [eventspace_name] + self.record_variables
+        CodeRunner.__init__(self, group=self, code=code, template='spikemonitor',
+                            name=None,  # The name has already been initialized
+                            clock=source.clock, when=when,
+                            order=order, needed_variables=needed_variables,
+                            template_kwds=template_kwds)
+
         self.variables.create_clock_variables(self._clock,
                                               prefix='_clock_')
+
+        self.add_dependency(source)
         self._enable_group_attributes()
 
     @property
@@ -97,6 +155,8 @@ class EventMonitor(Group, CodeRunner):
     def resize(self, new_size):
         self.variables['i'].resize(new_size)
         self.variables['t'].resize(new_size)
+        for variable in self.record_variables:
+            self.variables[variable].resize(new_size)
 
     def __len__(self):
         return self._N
@@ -121,9 +181,107 @@ class EventMonitor(Group, CodeRunner):
         '''
         return self.i, self.t_
 
+    def _values_dict(self, first_pos, sort_indices, used_indices, var):
+        sorted_values = self.state(var, use_units=False)[sort_indices]
+        dim = self.variables[var].unit.dim
+        event_values = {}
+        current_pos = 0  # position in the all_indices array
+        for idx in xrange(len(self.source)):
+            if current_pos < len(used_indices) and used_indices[current_pos] == idx:
+                if current_pos < len(used_indices) - 1:
+                    event_values[idx] = Quantity(sorted_values[
+                                                 first_pos[current_pos]:
+                                                 first_pos[current_pos + 1]],
+                                                 dim=dim, copy=False)
+                else:
+                    event_values[idx] = Quantity(
+                        sorted_values[first_pos[current_pos]:],
+                        dim=dim, copy=False)
+                current_pos += 1
+            else:
+                event_values[idx] = Quantity([], dim=dim)
+        return event_values
+
+    def values(self, var):
+        '''
+        Return a dictionary mapping neuron indices to arrays of variable values
+        at the time of the events (sorted by time).
+        Parameters
+        ----------
+        var : str
+            The name of the variable.
+
+        Returns
+        -------
+        values : dict
+            Dictionary mapping each neuron index to an array of variable
+            values at the time of the events
+
+        Examples
+        --------
+        >>> from brian2 import *
+        >>> G = NeuronGroup(2, """dv/dt = 100*Hz : 1
+        ...                       v_th : 1""", threshold='v>v_th', reset='v=0')
+        >>> G.v_th = [0.5, 1]
+        >>> mon = EventMonitor(G, event='spike', variables='v')
+        >>> run(20*ms)
+        >>> v_values = mon.values('v')
+        >>> v_values[0]
+        array([ 0.5,  0.5,  0.5,  0.5])
+        >>> v_values[1]
+        array([ 1.,  1.])
+        '''
+        indices = self.i[:]
+        sort_indices = np.argsort(indices)
+        used_indices, first_pos = np.unique(self.i[:][sort_indices],
+                                            return_index=True)
+        return self._values_dict(first_pos, sort_indices, used_indices, var)
+
+    def all_values(self):
+        '''
+        Return a dictionary mapping recorded variable names (including ``t``)
+        to a dictionary mapping neuron indices to arrays of variable values at
+        the time of the events (sorted by time). This is equivalent to (but more
+        efficient than) calling `values` for each variable and storing the
+        result in a dictionary.
+
+        Returns
+        -------
+        all_values : dict
+            Dictionary mapping variable names to dictionaries which themselves
+            are mapping neuron indicies to arrays of variable values at the
+            time of the events.
+
+        Examples
+        --------
+        >>> from brian2 import *
+        >>> G = NeuronGroup(2, """dv/dt = 100*Hz : 1
+        ...                       v_th : 1""", threshold='v>v_th', reset='v=0')
+        >>> G.v_th = [0.5, 1]
+        >>> mon = EventMonitor(G, event='spike', variables='v')
+        >>> run(20*ms)
+        >>> all_values = mon.all_values()
+        >>> all_values['t'][0]
+        array([  4.9,   9.9,  14.9,  19.9]) * msecond
+        >>> all_values['v'][0]
+        array([ 0.5,  0.5,  0.5,  0.5])
+        '''
+        indices = self.i[:]
+        sort_indices = np.argsort(indices)
+        used_indices, first_pos = np.unique(self.i[:][sort_indices],
+                                            return_index=True)
+        all_values_dict = {}
+        for varname in ['t'] + self.record_variables:
+            all_values_dict[varname] = self._values_dict(first_pos,
+                                                         sort_indices,
+                                                         used_indices,
+                                                         varname)
+        return all_values_dict
+
     def event_trains(self):
         '''
         Return a dictionary mapping event indices to arrays of event times.
+        Equivalent to calling ``values('t')``.
 
         Returns
         -------
@@ -135,27 +293,7 @@ class EventMonitor(Group, CodeRunner):
         --------
         SpikeMonitor.spike_trains
         '''
-        indices = self.i[:]
-        sort_indices = np.argsort(indices)
-        used_indices, first_pos = np.unique(self.i[:][sort_indices],
-                                            return_index=True)
-        sorted_times = self.t_[:][sort_indices]
-
-        event_times = {}
-        current_pos = 0  # position in the all_indices array
-        for idx in xrange(len(self.source)):
-            if current_pos < len(used_indices) and used_indices[current_pos] == idx:
-                if current_pos < len(used_indices)-1:
-                    event_times[idx] = Quantity(sorted_times[first_pos[current_pos]:first_pos[current_pos+1]],
-                                                dim=second.dim, copy=False)
-                else:
-                    event_times[idx] = Quantity(sorted_times[first_pos[current_pos]:],
-                                                dim=second.dim, copy=False)
-                current_pos += 1
-            else:
-                event_times[idx] = Quantity([], dim=second.dim)
-
-        return event_times
+        return self.values('t')
 
     @property
     def num_events(self):
@@ -178,7 +316,8 @@ class SpikeMonitor(EventMonitor):
     the attributes `~SpikeMonitor.i` and `~SpikeMonitor.t` store all the indices
     and spike times, respectively. Alternatively, you can get a dictionary
     mapping neuron indices to spike trains, by calling the `spike_trains`
-    method.
+    method. If you record additional variables with the ``variables`` argument,
+    these variables can be accessed by their name (see Examples).
 
     Parameters
     ----------
@@ -187,12 +326,16 @@ class SpikeMonitor(EventMonitor):
     record : bool
         Whether or not to record each spike in `i` and `t` (the `count` will
         always be recorded).
+    variables : str or sequence of str, optional
+        Which variables to record at the time of the spike (in addition to the
+        index of the neuron). Can be the name of a variable or a list of names.
     when : str, optional
-        When to record the spikes, by default records spikes in the slot
-        ``'end'``.
+        When to record the events, by default records events in the same slot
+        where the event is emitted.
     order : int, optional
         The priority of of this group for operations occurring at the same time
-        step and in the same scheduling slot. Defaults to 0.
+        step and in the same scheduling slot. Defaults to the order where the
+        event is emitted + 1, i.e. it will be recorded directly afterwards.
     name : str, optional
         A unique name for the object, otherwise will use
         ``source.name+'_spikemonitor_0'``, etc.
@@ -204,17 +347,30 @@ class SpikeMonitor(EventMonitor):
     >>> from brian2 import *
     >>> spikes = SpikeGeneratorGroup(3, [0, 1, 2], [0, 1, 2]*ms)
     >>> spike_mon = SpikeMonitor(spikes)
-    >>> run(3*ms)
+    >>> net = Network(spikes, spike_mon)
+    >>> net.run(3*ms)
     >>> print(spike_mon.i[:])
     [0 1 2]
     >>> print(spike_mon.t[:])
     [ 0.  1.  2.] ms
     >>> print(spike_mon.t_[:])
     [ 0.     0.001  0.002]
+    >>> G = NeuronGroup(1, """dv/dt = (1 - v)/(10*ms) : 1
+    ...                       dv_th/dt = (0.5 - v_th)/(20*ms) : 1""",
+    ...                 threshold='v>v_th',
+    ...                 reset='v = 0; v_th += 0.1')
+    >>> crossings = SpikeMonitor(G, variables='v', name='crossings')
+    >>> net = Network(G, crossings)
+    >>> net.run(10*ms)
+    >>> crossings.t
+    <crossings.t: array([ 0. ,  1.4,  4.6,  9.7]) * msecond>
+    >>> crossings.v
+    <crossings.v: array([ 0.00995017,  0.13064176,  0.27385096,  0.39950442])>
     '''
-    def __init__(self, source, when='end', order=0,
+    def __init__(self, source, variables=None, when=None, order=None,
              name='spikemonitor*', codeobj_class=None):
-        super(SpikeMonitor, self).__init__(source, event='spike', when=when,
+        super(SpikeMonitor, self).__init__(source, event='spike',
+                                           variables=variables, when=when,
                                            order=order, name=name,
                                            codeobj_class=codeobj_class)
 
@@ -224,6 +380,10 @@ class SpikeMonitor(EventMonitor):
         Returns the total number of recorded spikes.
         '''
         return self.num_events
+
+    # We "re-implement" the following functions only to get more specific
+    # doc strings (and to make sure that the methods are included in the
+    # reference documentation for SpikeMonitor).
 
     def spike_trains(self):
         '''
@@ -246,6 +406,69 @@ class SpikeMonitor(EventMonitor):
         array([ 1.]) * msecond
         '''
         return self.event_trains()
+
+    def values(self, var):
+        '''
+        Return a dictionary mapping neuron indices to arrays of variable values
+        at the time of the spikes (sorted by time).
+
+        Parameters
+        ----------
+        var : str
+            The name of the variable.
+
+        Returns
+        -------
+        values : dict
+            Dictionary mapping each neuron index to an array of variable
+            values at the time of the spikes.
+
+        Examples
+        --------
+        >>> from brian2 import *
+        >>> G = NeuronGroup(2, """dv/dt = 100*Hz : 1
+        ...                       v_th : 1""", threshold='v>v_th', reset='v=0')
+        >>> G.v_th = [0.5, 1]
+        >>> mon = SpikeMonitor(G, variables='v')
+        >>> run(20*ms)
+        >>> v_values = mon.values('v')
+        >>> v_values[0]
+        array([ 0.5,  0.5,  0.5,  0.5])
+        >>> v_values[1]
+        array([ 1.,  1.])
+        '''
+        return super(SpikeMonitor, self).values(var)
+
+    def all_values(self):
+        '''
+        Return a dictionary mapping recorded variable names (including ``t``)
+        to a dictionary mapping neuron indices to arrays of variable values at
+        the time of the spikes (sorted by time). This is equivalent to (but more
+        efficient than) calling `values` for each variable and storing the
+        result in a dictionary.
+
+        Returns
+        -------
+        all_values : dict
+            Dictionary mapping variable names to dictionaries which themselves
+            are mapping neuron indicies to arrays of variable values at the
+            time of the spikes.
+
+        Examples
+        --------
+        >>> from brian2 import *
+        >>> G = NeuronGroup(2, """dv/dt = 100*Hz : 1
+        ...                       v_th : 1""", threshold='v>v_th', reset='v=0')
+        >>> G.v_th = [0.5, 1]
+        >>> mon = SpikeMonitor(G, variables='v')
+        >>> run(20*ms)
+        >>> all_values = mon.all_values()
+        >>> all_values['t'][0]
+        array([  4.9,   9.9,  14.9,  19.9]) * msecond
+        >>> all_values['v'][0]
+        array([ 0.5,  0.5,  0.5,  0.5])
+        '''
+        return super(SpikeMonitor, self).all_values()
 
     def __repr__(self):
         description = '<{classname}, recording from {source}>'